requirements.txt
apscheduler>=3.3.0,<4.0
aws-requests-auth>=0.3.0
sortedcontainers>=2.2.2
boto3>=1.4.4
cffi>=1.11.5
croniter>=0.3.16
elasticsearch==7.0.0
envparse>=0.2.0
exotel>=0.1.3
Jinja2==3.0.1
jira>=2.0.0
jsonschema>=3.0.2
prison>=0.1.2
prometheus_client>=0.10.1
py-zabbix>=1.1.3
python-dateutil>=2.6.0,<2.9.0
PyYAML>=5.1
requests>=2.10.0
stomp.py>=4.1.17
texttable>=0.8.8
statsd-tags==3.2.1.post1
twilio>=6.0.0,<6.58
tzlocal<3.0
pyyaml-include>=1.2.post2